requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon
spamwatch
alphabet_detector
youtube-dl
tswift
gTTS
google-api-python-client
